### 2020-07-02 #2421 @whedon
[PRE REVIEW]: GeoBO: Python package for Multi-Objective Bayesian Optimisation and Joint Inversion in Geosciences
**Submitting author:** @sebhaan (<a href="http://orcid.org/0000-0002-5994-5637">Sebastian Haan</a>)
**Repository:** <a href="https://github.com/sebhaan/geobo" target ="_blank">https://github.com/sebhaan/geobo</a>
**Version:** v1.0.0
