### 2020-07-02 #2421 @jedbrown
:wave: @sebhaan Could you explain a bit more about related software. Perhaps https://hippylib.github.io/, for example? This would be a good thing to add to your README/documentation as well as the paper.
