### 2020-07-04 1401f2c
Add related inversion methods and references to paper
