+++ paper/paper.md
@@ -24,28 +24,33 @@ A critical decision process in geophysical data collection is how to efficiently
 
 ``GeoBO`` is a Python package that solves both, Multi-Objective Optimisation and Joint Inversion, within one probabilistic framework: from prior selection and input, data fusion and inversion, to the final sensor optimisation and real world model output. Fig.1 shows a graphical overview model of ``GeoBO``; the two main process steps are summarized in the following. 
 
-First, ``GeoBO`` jointly solves multi-linear forward models of 2D survey data (e.g. magnetics and gravity) to 3D-geophysical properties using Gaussian Process kernels (see e.g. [@Rasmussen:2006], [@Melkumyan:2009]), without the requirement of prior geological knowlegde. In addition, sparse drillhole measurements can be taken into account.  This joint inversion provides a better constrained combined solution of multiple, distinct sensor types, rather than taking their individual solutions. One of the key features is to consider all cross-variances between geophysical properties by solving simultaneously multiple forward models ([@Melkumyan:2011], [@Reid:2013]). Thus, this approach can take full advantage of additional sensor information to reveal properties of the reconstructed geophysical model at higher accuracy. By using non-parametric Gaussian Process priors, the solution of the inverse problem provides a complete posterior distribution for all geophysical properties which are described by their mean and variance value at each location (voxel).
+First, ``GeoBO`` jointly solves multi-linear forward models of 2D survey data (e.g. magnetics and gravity) to 3D-geophysical properties using Gaussian Process kernels [see e.g. @Rasmussen:2006; @Melkumyan:2009], without the requirement of prior geological knowledge. In addition, sparse drillhole measurements can be taken into account. One of the key features is to consider correlations  between geophysical properties by solving simultaneously multiple forward models [@Melkumyan:2011; @Reid:2013]. Thus, this approach provides a better constrained combined solution of multiple, distinct sensor types, rather than taking their individual solutions. By using non-parametric Gaussian Process priors, the solution of the inverse problem provides a complete posterior distribution for all geophysical properties which are described by their mean and variance value at each location (voxel). While this inversion method is limited to linear forward models, the reconstructed posterior distribution of the geophysical properties can be further refined using site-specific knowledge or more complex inversion methods for sampling non-linear forward models, such as hIPPYlib [@Hippylib], Obsidian [@Obsidian], or GemPy [@GemPy].
 
-In a second step, ``GeoBO`` allows the user to define objectives in an aquisitian function ([@Brochu:2010]) which typically has to balance between a) exploration, i.e., querying points that maximise the information gain and minimize the uncertainty of a geophysical site model, b) exploitation, i.e. querying points that maximise the reward (e.g. concentrating search in the vicinity locations with high value such as minerals), and c) minimize the number of samples given a cost function for any new measurement (e.g. costly drill-cores). The reconstructed 3D model output of the joint inversion model is then used to query for the next most promising point based on the aquisitian function, which guides the search for a user-defined optimum.
+In a second step, ``GeoBO`` allows the user to define objectives in an acquisition function [@Brochu:2010] which typically has to balance between a) exploration, i.e., querying points that maximise the information gain and minimize the uncertainty of a geophysical site model, b) exploitation, i.e. querying points that maximise the reward (e.g. concentrating search in the vicinity locations with high value such as minerals), and c) minimize the number of samples given a cost function for any new measurement (e.g. costly drill-cores). The reconstructed 3D model output of the joint inversion model is then used to query for the next most promising point based on the aquisitian function, which guides the search for a user-defined optimum.
 
[…]
