### 2020-07-04 eac3c8a
Add references to paper.bib
