+++ paper/paper.bib
@@ -55,3 +55,33 @@
   journal={Geophysics, submitted},
   year={2020}
 }
+
+@article{Hippylib,
+  title = "{hIPPYlib: an Extensible Software Framework for Large-scale Deterministic and Bayesian Inverse Problems}",
+  author = {Villa, U. and Petra, N. and Ghattas, O.},
+  journal = {Journal of Open Source Software},
+  volume = {3},
+  number = {30},
+  page = {940},
+  doi  = {10.21105/joss.00940},
+  year = {2018}
+}
+}
+
+@article{Obsidian,
+  title={Efficiency and robustness in Monte Carlo sampling for 3-D geophysical inversions with Obsidian v0. 1.2: setting up for success},
+  author={Scalzo, Richard and Kohn, David and Olierook, Hugo and Houseman, Gregory and Chandra, Rohitash and Girolami, Mark and Cripps, Sally},
+  journal={Geoscientific Model Development},
+  volume={12},
+  number={7},
+  pages={2941--2960},
+  year={2019},
+  publisher={Copernicus GmbH}
+}
+
+@article{GemPy,
+  title={GemPy 1.0: open-source stochastic geological modeling and inversion},
+  author={de la Varga, Miguel and Schaaf, Alexander and Wellmann, Florian},
+  journal={Geoscientific Model Development},
+  year={2019}
+}
