### 2020-07-04 #2421 @sebhaan
> 👋 @sebhaan Could you explain a bit more about related software. Perhaps https://hippylib.github.io/, for example? This would be a good thing to add to your README/documentation as well as the paper.

@jedbrown Have added a new section about related software (hippylib, Obsidian, and GemPy) to the README and three more references to these bayesian inversion tools in the paper.
