### 2020-08-19 #2421 @hugoledoux
I have already asked 2 other persons (on the JOSS reviewers list).

I just saw it was close to your expertise and thought that would be an easy way to recruit the first reviewer! But I understand, let's wait then.
