### 2020-09-18 #2421 @hugoledoux
Would @uvilla or @npetra would be available for this review? I didn't get any response from the 2 on the JOSS list, unfortunately.
