### 2020-09-19 #2421 @uvilla
Hi Hugo,
 I’m on a family leave until the end of the semester, and unfortunately I
won’t be able to serve as a reviewer for Joss this time around.

I’m looking forward to serve in a future occasion.

Kind regards,

Umberto

On Fri, Sep 18, 2020 at 1:38 AM Hugo Ledoux <notifications@github.com>
wrote:

>
>
> Would @uvilla <https://github.com/uvilla> or @npetra
> <https://github.com/npetra> would be available for this review? I didn't
> get any response from the 2 on the JOSS list, unfortunately.
>
>
>
>
