### 2020-09-22 #2421 @hugoledoux
@npetra great. We try to do the reviews <1mo, would that be possible?

I don't know if you've done reviews for JOSS before? If not, you should know that the review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository
and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Editors and reviewers post comments on the Review issue, and authors respond to the comments
and improve their submission until acceptance (or withdrawal, if they feel unable to
satisfy the review).

Let me know if possible, if not, if you could please propose alternatives that would be great; their github username here (without the "@" so that they don't get automatically the notification).
