### 2020-09-22 #2421 @npetra
@hugoledoux 1 month should be fine. Thanks for the details. This is the first time I am reviewing for JOSS, will take a look at the review criteria and get back to you if I have questions.
