### 2020-09-22 #2421 @hugoledoux
@npetra super. The actual review is done in another issue, you'll be notified. At the top of the issue there are instructions about what you should do. It's a bit overwhelming at first, so don't hesitate to ask me questions.

You are allowed--and expected!--to interact directly with the authors. You can say "hey this and that doesn't work" and we expect them to fix those. It's a dialogue that goes back and forth.
