### 2020-09-22 #2690 @whedon
[REVIEW]: GeoBO: Python package for Multi-Objective Bayesian Optimisation and Joint Inversion in Geosciences
**Submitting author:** @sebhaan (<a href="http://orcid.org/0000-0002-5994-5637">Sebastian Haan</a>)
**Repository:** <a href="https://github.com/sebhaan/geobo" target ="_blank">https://github.com/sebhaan/geobo</a>
**Version:** v1.0.0

@npetra and @sgkang, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
