### 2020-10-28 #2690 @npetra
Hi @sebhaan (& co),

GeoBO seems to be a useful and great software package for the community. Thanks for working on it and for sharing it with the community.

Here are my comments:

- [ ] The package is successfully installed by “python setup.py install”. However, it seems that the package is not registered in PyPI, and cannot be installed through ‘pip3’.

- [ ] Is the software extensible to other physical problems (different linear forward models, e.g., elasticity models)? If it is, how intrusive such extension will be.

- [ ] Is it possible to use a user-defined prior function?

- [ ] What does colors mean in the result figures, for example, in the figure ‘newdrill_proposals.png’?

- [ ] It would begreat to provide a clear explanation of the generated files and of the setting file (the authors provide some comments for input variables in the setting file, but it may not be enough to fully understand what the variables mean. For example, it is hard for me to get what the constants (c_G, c_SI_TO_MILLIAGALS, …) in the setting file mean).

**Software/Codes**

- [ ] The copyright information is not complete and consistent throughout the code. There seems to be one single author for GeoBO. Perhaps make this clear in the Authors list/discussion.

- [ ] It would be great if the authors could also provide a Docker image. This would ease up the installation struggles users might have with the installation of the software and its dependencies.

- [ ] There is no CONTRIBUTION file or section in README file for a guideline for third parties wishing to contribute to the software or to report issues or problems with the software.

**Paper:**

- [ ] The author list is missing from the paper.

- [ ] From the second setup it sounds like the authors also tackle optimal experimental design problems. If so, perhaps this should be clearly stated and a few references given.
