### 2020-10-29 #2690 @sebhaan
Hi @npetra,

thank you for your very helpful comments and taking your time to review this package.
I will have some more time to work on it next week and will come back to you soon with the details for each point.
