### 2020-11-02 #2690 @sgkang
Dear @sebhaan et al.,

This is an interesting work, and I am glad to have a chance to review this article.
When reading this article, I got a rough idea what authors are trying to accomplish with `geobo`, but still it was not clear to me. So, I was hoping to get better idea from examples. But, it ended up generating images and other outputs, which were hard for me to decipher what they mean. I would recommend having tutorial notebooks, or better form of documentation (e.g., sphinx).

My detailed comments are following:

- [ ] Pypi version did not work, so I had to clone and run `python setup.py install`
- [ ] There are test examples, but no tests. I was not sure how to quantify the performance of the code base. For instance, for geophysical simulation problem, I would expect to have comparison with analytic solution.

- [ ] I was not sure "the statement of need" is clear. For instance, why do I need to use `geobo` rather than using conventional mag. and gravity inversion methods (e.g. Li and Oldenburg, 1998)?
- [ ] There is an example, and it runs. But, I was not sure that I could understand what the output means. There should be some description about the outputs.
- [ ] I thought documentation is sparse. For the current form, I am not sure that I can run other examples than provided in the repository.
- [ ] I have not found automated tests (e.g., travis-ci), and I recommend authors to generate test cases (not examples, but testing performance of the codes)

- [ ] I was not sure what are the problems (or applications) that authors are having in mind. I felt it is designed to solve mining situations. Also, I was not sure who are the target audiences. So, these should be clarified.
- [ ] Comparison with other software was missing. If the goal is imaging the distribution of physical properties, I recommend suggesting a typical 3D potential field inversion method (e.g. Li and Oldenburg, 1998).
- [ ] Thought interesting that authors solves an optimal survey design problem. But, I was a bit lost how that problem is connected with joint inversion; it seems the connections are sensor data & fusion and world model with uncertainty (from Figure 1), but I was not sure what they mean.
