### 2020-11-08 a4bd0a3
Add 3d cube visualisation of magsus
