### 2020-11-08 15c785b
Add more sections to README (Options, Customization and Result files)
