+++ README.md
@@ -14,7 +14,11 @@
 - [Usage and Settings](#usage-and-settings) 
 - [Examples and Tests](#examples)
   - [Synthetic Models](#synthetic-models)
-  - [Drillcore Test Example](#drillcore-estt-example)
+  - [Drillcore Test Example](#drillcore-test-example)
+  - [Results and Output Files](#results-and-output-files)
+- [Options and Customization](#options-and-customization)
+  - [Custom Linear Forward Models](#custom-linear-forward-models)
+  - [Gaussian Process Kernel Functions][#gaussian-process-kernel-functions]
 - [Literature](#literature)
 - [Related Software](#related-software)
 - [Attribution and Acknowledgments](#attribution-and-acknowledgements)
@@ -29,7 +33,8 @@ to evaluate, do not have a closed-form (e.g. black-boxfunctions), or have no acc
 
 ### Acquisition function 
 
-The key of BO is the acquisition function, which typically has to balance between 
+The key of BO is the acquisition function, which typically has to balance between: 
+
 a) exploration, i.e., querying points that maximise the information gain and minimize the uncertainty of a model
 b) exploitation, i.e. querying points that maximise the reward (e.g. concentrating search in the
 vicinity locations with high value such as minerals)
@@ -46,7 +51,7 @@ Forward models transform the localized measurement of a remote sensor grid into 
 
 GeoBO's probabilistic framework includes all steps from  prior selection, data fusion and inversion, to sensor optimisation and real world model output. The main functionalities of GeoBO are summarised in the following:
 
- - Joint probabilistic inversion tool by solving simultaneously multi-linear forward models (e.g. gravity, magnetics) using cross-variances between geophysical properties (cross-variance terms can be specified by user)
+ - Joint probabilistic inversion tool by solving simultaneously multi-linear forward models (e.g. gravity, magnetics) using cross-variances between geophysical properties (cross-variance terms can be specified by user). 
  - Output 1: Generation of cubes and computation of complete posterior distribution for all geophysical properties (described by their mean and variance value at each location (cubecell aka voxel). 
  - Output 2: Generation of ranked proposal list for new most promising drillcores based on global optimisation of acquisition function
  - Templates for acquisition function to use in Bayesian Optimisation
@@ -60,6 +65,7 @@ Other features are:
  - Generation of 2D/3D visualisation plots of reconstructed cubes and survey data
  - 3D Cube export in VTK format (for subsequent analysis, e.g., in Python or with ParaView)
  - Options to include any pre-existing drillcore data 
+ - Included linear forward models: density-to-gravity and magnetic susceptibility-to-magnetic field; custom linear forward models can be added (see [Custom Linear Forward Models](#custom-linear-forward-models))
  - Library of Gaussian Process (GP) kernels including sparse GP kernels
  - Flexible settings for any cube geometry and resolution
  - (Optional) Optimization of GP hyperparameters and cross-correlation coefficients via computation of marginal GP likelihood
@@ -167,9 +173,6 @@ cd geobo/
 python main.py tests/settings_example1.yaml 
 ```
 
-The output results include the generated reconstructed density and magnetic susceptibility cubes and their corresponding uncertainty cubes, visualisations of original survey data and reconstructed properties, and list of new drillcore proposals.
-The output figure 'newdrill_proposals.png' shows the location of the already existing drills (black points), proposed new drill positions (white), and  the best new drill location (red). 
-
 
 ### Drillcore Test Example
 
@@ -182,9 +185,60 @@ python main.py tests/settings_example2.yaml
 and creates the reconstructed density and magnetic susceptibility cubes, uncertainty cubes
 
 
+### Results and Output Files
+
[…]
