### 2020-11-08 20eb9ba
Minor update in documentatiopn to sensormodel functions
