### 2020-11-09 ab3f42c
Add consistent copyright and author information.
