+++ geobo/inversion.py
@@ -1,7 +1,14 @@
 """
 Script for running inversion and reconstructing 3D cubes from 2D sensor data using Gaussina processes
 
-Author: Sebastian Haan
+Copyright 2020 Sebastian Haan
+
+This file is part of GeoBO.
+
+GeoBO is free software made available under the AGPL License. 
+For details see the LICENSE file.
+
+@author: Sebastian Haan
 """
 
 import numpy as np
