+++ geobo/cubeshow.py
@@ -3,8 +3,14 @@ Framework for plotting 3D volumetric data (3D cube) as voxel grid.
 
 Includes export of datacube as vtk file
 
-Author: Sebastian Haan
+Copyright 2020 Sebastian Haan
 
+This file is part of GeoBO.
+
+GeoBO is free software made available under the AGPL License. 
+For details see the LICENSE file.
+
+@author: Sebastian Haan
 """
 
 import os 
