### 2020-11-10 d56a017
Removed latex equations since Github still doesn't support equations natively (seriously?)
