### 2020-11-13 c98e74f
Add CONTRIBUTING.md guideline file
