### 2020-11-14 edf5d10
Update paper with review comments
