### 2020-11-15 f4733bb
Update copyright and license description.
