### 2020-11-15 9269490
added a few more descriptions to settings
