### 2020-11-15 #2690 @sebhaan
Dear @sgkang

thank you for taking your to review this package and your helpful comments. I've updated the code, paper, and documentation accordingly, but still have to add some more automated testing. Please see below inline for the detailed reply.

> Dear @sebhaan et al.,
>
> This is an interesting work, and I am glad to have a chance to review this article.
> When reading this article, I got a rough idea what authors are trying to accomplish with `geobo`, but still it was not clear to me. So, I was hoping to get better idea from examples. But, it ended up generating images and other outputs, which were hard for me to decipher what they mean. I would recommend having tutorial notebooks, or better form of documentation (e.g., sphinx).
>
> My detailed comments are following:
>
> ### Functionality
> * [ ]  Pypi version did not work, so I had to clone and run `python setup.py install`

Re:  It is now registered on PyPi (https://pypi.org/project/geobo/0.1.0/) and can be installed with pip.

> * [ ]  There are test examples, but no tests. I was not sure how to quantify the performance of the code base. For instance, for geophysical simulation problem, I would expect to have comparison with analytic solution.

Re: Will add some more automated test (likely with pytest). I presume the analytic solution refers to the forward model testing.  In principle one could include a test of the forward model analytically for one voxel at a given distance from sensor on surface. GeoBO has been tested using the following approach by "closing the circle test" between forward modeling and inversion. Assuming the forward model is correct (e.g. via one voxel test), the simulation creates first synthetic 3D subsurface properties, then computes the corresponding sensor measurements on the surface using the deterministic forward model, and then can finally test if the inversion can reconstruct the synthetic 3D subsurface
properties again.

>
> ### Documentation
> * [ ]  I was not sure "the statement of need" is clear. For instance, why do I need to use `geobo` rather than using conventional mag. and gravity inversion methods (e.g. Li and Oldenburg, 1998)?

 Re: As described in the updated paper there are two main advantages in comparison to traditional inversion methods:
 First,﻿one key advantage of this method is to solve simultaneously multiple sensor measurements by taking into account the correlations between distinct geological aspect, e.g., density and magnetic susceptibility. This joint optimization has the advantage to find the best possible solution over the complete parameter space.
Another practical advantage of this probabilistic approach is that predictions are described by posterior distributions for each location (voxel or cube cell), which quantify the uncertainty in the predictions and their credible intervals. Most other work in the field of joint inversion (e.g. Li and Oldenburg, 1998) only report point estimates of the quantities of interest.
This ability to propagate consistently uncertainties from input to output space under the Bayesian inversion formalism is an essential requirement for the subsequent Bayesian optimisation to search for new optimal measurement locations (since the acquisition function needs the predicted uncertainty cube as input).

> * [ ]  There is an example, and it runs. But, I was not sure that I could understand what the output means. There should be some description about the outputs.

Re: Added to README file and paper an additional section (Results and Output Files) to describe the output.

> * [ ]  I thought documentation is sparse. For the current form, I am not sure that I can run other examples than provided in the repository.

Re: Added a couple of new sections in the README file to describe better the user and customization options such as "Custom Linear Forward Models" , which should help to add any new forward model to the module sensormodel.py. Also more details about Gaussian Process Kernel Functions and Bayesian Optimisation Options have been added to README.
For more mathematical details, please see the paper "Multi-Objective Bayesian Optimisation and Joint Inversion for Active Sensor Fusion" (https://arxiv.org/abs/2010.05386)

> * [ ]  I have not found automated tests (e.g., travis-ci), and I recommend authors to generate test cases (not examples, but testing performance of the codes)

Re: Yes, it currently doesn't has any automated test via pytest or travis given the complexity of user options, except for running the test examples. On the current To-Do list is  adding automated testing of these and other critical functions to the code. Also testing the performance of the code via code profilers can be added.

> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Re: Added file CONTRIBUTION.md with instructions

>
> ### Software paper
> * [ ]  I was not sure what are the problems (or applications) that authors are having in mind. I felt it is designed to solve mining situations. Also, I was not sure who are the target audiences. So, these should be clarified.

Re: Added more description to paper. While the current implementation generates 3D geophysical properties based on gravity and magnetic inversion and searching for optimal new drillcore measurements, in principle the same model can be applied to a wider range of sensor fusion problems and allocation problems. For instance:  What is the optimal mixture and placement of different sensors?  Where to sample if the cost function is incomplete or if there are large uncertainties in future total budget allocations? How to position sensor grids over time if the model state is dynamic or has a moving target?
Moreover GeoBO is a data-driven approach and uses Gaussian Process method for solving the inversion problem. The advantages of the Gaussian Process method is that it is tractable, generates posterior distributions, can predict at any resolution or scale, and can be optimised via marginal log likelihood.  An alternative to Gaussian Process method could be using sampling methods (e.g. via MCMC or variational inference) but they are computationally very expensive given their need to sample over an almost infinite range of possible models, and are therefore in practice typically constrained by certain assumptions about the underlying geological formation and  prior parameterisation.

> * [ ]  Comparison with other software was missing. If the goal is imaging the distribution of physical properties, I recommend suggesting a typical 3D potential field inversion method (e.g. Li and Oldenburg, 1998).

Re: See comment above (the reply on "the statement of need"). Li and Oldenburg is not a joint inversion and does not generate posterior distributions (uncertainty cubes).

> * [ ]  Thought interesting that authors solves an optimal survey design problem. But, I was a bit lost how that problem is connected with joint inversion; it seems the connections are sensor data & fusion and world model with uncertainty (from Figure 1), but I was not sure what they mean.

Re: Bayesian Optimisation (BO) does not necessarily need a joint inversion problem but requires that the inversion method (or any 3D surrogate world modeling) - no matter whether it is a joint or simple - generates cubes with location dependent uncertainty predictions (heteroscedastic noise). The objective function in BO requires the information about the uncertainty to find new locations that increase information gain (reduce uncertainty). The Gaussian Process inversion method produces basically a probabilistic surrogate model of the 3D world properties.
Note that BO is not to the same as the "Design of Experiment" method that aims to search for an effective initial parameter (experiment variables) configuration typically BEFORE any measurement is taken (and before any model evaluation is possible). BO aims to improve a model typically AFTER some initial measurements are taken, which allows us to build a surrogate world model.
