### 2020-11-18 81fecb6
Add more documentation to README and paper.md
