+++ paper/paper.bib
@@ -52,7 +52,7 @@
 @article{Haan:2020,
   title={Multi-Objective Bayesian Optimisation and Joint Inversion for Active Sensor Fusion},
   author={Sebastian Haan, Fabio Ramos, Dietmar Muller},
-  journal={Geophysics, submitted},
+  journal={Geophysics, 86(1), pp.1-78.},
   year={2020}
 }
 
@@ -79,9 +79,51 @@
   publisher={Copernicus GmbH}
 }
 
+@article{Gallardo:2003,
+  title={{A versatile algorithm for joint 3D inversion of gravity and magnetic data}},
+  author={Gallardo-Delgado, Luis A and P{\'e}rez-Flores, Marco Antonio and G{\'o}mez-Trevi{\~n}o, Enrique},
+  journal={Geophysics},
+  volume={68},
+  number={3},
+  pages={949--959},
+  year={2003},
+  publisher={Society of Exploration Geophysicists}
+}
+
 @article{GemPy,
   title={GemPy 1.0: open-source stochastic geological modeling and inversion},
   author={de la Varga, Miguel and Schaaf, Alexander and Wellmann, Florian},
   journal={Geoscientific Model Development},
   year={2019}
+}
+
+@article{Mockus:1975,
+  title={{On Bayesian methods of optimization}},
+  author={Mockus, J},
+  journal={Towards Global Optimization},
+  pages={166--181},
+  year={1975},
+  publisher={New York: North-Holland}
+}
+
+@article{Moorkamp:2011,
+  title={{A framework for 3-D joint inversion of MT, gravity and seismic refraction data}},
+  author={Moorkamp, Max and Heincke, Bj{\"o}rn and Jegen, Marion and Roberts, Alan W and Hobbs, Richard W},
+  journal={Geophysical Journal International},
+  volume={184},
+  number={1},
+  pages={477--493},
+  year={2011},
+  publisher={Blackwell Publishing Ltd Oxford, UK}
+}
+
+@article{Zeyen:1993,
+  title={{3-D joint inversion of magnetic and gravimetric data with a priori information}},
+  author={Zeyen, Hermann and Pous, Jaume},
+  journal={Geophysical Journal International},
+  volume={112},
+  number={2},
+  pages={244--256},
+  year={1993},
+  publisher={Blackwell Publishing Ltd Oxford, UK}
 }
