### 2020-11-18 2173614
Revision of typo in magnetic suceptibility calculation
