### 2021-01-11 #2690 @hugoledoux
> @sebhaan The new install seems to work and all my comments have been addressed. This paper is good to go from my perspective.

thanks @npetra , could you check the list items above please if all is satisfactory?
