### 2021-01-17 #2690 @sgkang
Dear @sebhaan,

My comments are addressed, and the paper is looking good to me. Look forward to this paper is out the door!

Below are my suggests that I thought while I was reviewing your paper and code base.

1. If you can add few other open-source geophysical inversion packages that would be nice

> Cockett, R., Kang, S., Heagy, L. J., Pidlisecky, A., & Oldenburg, D. W. (2015). SimPEG: An open source framework for simulation and gradient based parameter estimation in geophysical applications. Computers & Geosciences, 85, 142–154. https://doi.org/http://dx.doi.org/10.1016/j.cageo.2015.09.015"""

> Rücker, C., Günther, T., & Wagner, F. M. (2017). {pyGIMLi}: An open-source library for modelling and inversion in geophysics. Computers and Geosciences, 109, 106–123. https://doi.org/10.1016/j.cageo.2017.07.011

2. Generalization of the code (?). Had a brief look at `Inversion` class, and it is tied to magnetic and gravity methods, so making the inversion class bit more general seems a reasonable step towards.

3. Testing of the code. Current form of testing, which runs a synthetic example and check the results seems okay at the moment. But, for a long run, I would recommend implementing more rigorous tests, more than testing with the synthetic example (e.g. expected property of your optimization scheme...). Further use the `unittest`, and automate the testing procedure such that you can readily deploy of continuous integration platform like travis.

Hope this helps,

All the best,

Seogi.
