### 2021-01-18 #2690 @sebhaan
Hi @sgkang,

thanks so much for your suggestions,  on the list now for the next version. The suggested inversion packages are very interesting as well and could potentially be included if they can provide predictive model uncertainties for the BO function.

Thanks again for the review,
Seb
