### 2021-01-18 a0f516f
Fix some errors, mostly related to the citations
