### 2021-01-18 #2690 @hugoledoux
ok @sebhaan it seems that the reviewers are now happy with the changes you made, and we can move towards the final acceptance.

I checked you PDF and above you can see that many DOI are missing, could you please add them and run again here `@whedon check references` please?

Also, the way you added references was sometimes wrong, so I fixed them (plus 2-3 minor cosmetic things) in that PR: https://github.com/sebhaan/geobo/pull/1
