### 2021-01-19 ce0fdb7
Add DOIs to bib references.
