+++ paper/paper.bib
@@ -53,6 +53,7 @@
   title={Multi-Objective Bayesian Optimisation and Joint Inversion for Active Sensor Fusion},
   author={Sebastian Haan, Fabio Ramos, Dietmar Muller},
   journal={Geophysics, 86(1), pp.1-78.},
+  doi = {10.1190/geo2019-0460.1},
   year={2020}
 }
 
@@ -75,6 +76,7 @@
   volume={12},
   number={7},
   pages={2941--2960},
+  doi={10.5194/gmd-12-2941-2019},
   year={2019},
   publisher={Copernicus GmbH}
 }
@@ -86,6 +88,7 @@
   volume={68},
   number={3},
   pages={949--959},
+  doi={10.1190/1.1581067},
   year={2003},
   publisher={Society of Exploration Geophysicists}
 }
@@ -94,6 +97,7 @@
   title={GemPy 1.0: open-source stochastic geological modeling and inversion},
   author={de la Varga, Miguel and Schaaf, Alexander and Wellmann, Florian},
   journal={Geoscientific Model Development},
+  doi={10.5194/gmd-12-1-2019},
   year={2019}
 }
 
@@ -113,17 +117,19 @@
   volume={184},
   number={1},
   pages={477--493},
+  doi={10.1111/j.1365-246X.2010.04856.x},
   year={2011},
   publisher={Blackwell Publishing Ltd Oxford, UK}
 }
 
 @article{Zeyen:1993,
-  title={{3-D joint inversion of magnetic and gravimetric data with a priori information}},
+  title={{3-D Joint Inversion of Magnetic and Gravimetric data with a priori Information}},
   author={Zeyen, Hermann and Pous, Jaume},
   journal={Geophysical Journal International},
   volume={112},
   number={2},
   pages={244--256},
+  doi={10.1111/j.1365-246x.1993.tb01452.x},
   year={1993},
   publisher={Blackwell Publishing Ltd Oxford, UK}
 }
