### 2021-01-19 f2529ea
Remove double parentheses in citations.
