+++ paper/paper.bib
@@ -50,7 +50,7 @@
 }
 
 @article{Haan:2020,
-  title={Multi-Objective Bayesian Optimisation and Joint Inversion for Active Sensor Fusion},
+  title={Multiobjective Bayesian optimisation and joint inversion for active sensor fusion},
   author={Sebastian Haan, Fabio Ramos, Dietmar Muller},
   journal={Geophysics, 86(1), pp.1-78.},
   doi = {10.1190/geo2019-0460.1},
