+++ paper/paper.md
@@ -1,0 +1,83 @@
+---
+title: 'GeoBO: Python package for Multi-Objective Bayesian Optimisation and Joint Inversion in Geosciences'
+tags:
+  - Python
+  - Geoscience
+  - Bayesian Optimisation
+  - Inversion
+  - Gaussian Processes
+authors:
+  - name: Sebastian Haan
+    orcid: 0000-0002-5994-5637
+    affiliation: "1"
+affiliations:
+  - name: Sydney Informatics Hub, The University of Sydney, Australia
+    index: 1
+date: 14 November 2020
+bibliography: paper.bib
+---
+<!-- pandoc -V geometry:margin=1in -V fontsize:11pt --filter pandoc-citeproc  -o paper.pdf paper.md -->
+<!--add "author: Sebastian Haan" to meta for standard pandoc conversion --> 
+
+
+# Summary
+
+A critical decision process in geophysical data collection is how to efficiently combine a variety of sensor types and where to collect new observations, in particular if measurements are very costly or resources are limited. This may include drill-core placements or the layout of a geophysical survey. Bayesian Optimisation (BO) solves this decision making problem by finding global optimal solutions given multiple, often distinct measurements and model uncertainties. One of the major challenges for applying BO in geoscience still lies in constructing a 3D model from sparse, typically 2D sensor data that can be computationally efficiently evaluated and that transforms the combined data and uncertainties from multiple sensor measurements coherently into a posterior function approximating the true model; this is also know as an inversion problem.
+
+``GeoBO`` is a Python package that solves both, Multi-Objective Optimisation and Joint Inversion, within one probabilistic framework: from prior selection and input, data fusion and inversion, to the final sensor optimisation and real world model output. 
+<!-- While the current implementation generates 3D geophysical properties based on gravity and magnetic inversion and searching for optimal new drillcore measurements, in principle, the same model can be applied to a wide range of sensor fusion problems and allocation problems, such as: How to position or activate sensors for quasi-linear inverse problems (e.g. seismic, tomography) or if the model state is dynamic?  Where to sample if the cost function is incomplete? -->
+Fig.1 shows a graphical overview model of ``GeoBO``; the two main process steps are summarized in thee following. 
+
+First, ``GeoBO`` jointly solves multi-linear forward models of 2D survey data (e.g. magnetics and gravity) to 3D-geophysical properties using Gaussian Process kernels [see, e.g., @Rasmussen:2006; @Melkumyan:2009], without the requirement of prior geological knowledge. In addition, sparse drillhole measurements can be taken into account. One of the key features is to consider correlations between geophysical properties by solving simultaneously multiple forward models [@Melkumyan:2011; @Reid:2013]. Thus, this approach provides a better constrained combined solution of multiple, distinct sensor types, rather than taking their individual solutions. Another practical advantage of this probabilistic approach is that predictions are described by posterior distributions for each location (voxel or cube cell), which quantify the uncertainty in the predictions and their credible intervals. Most other work in the field of joint inversion only report point estimates of the quantities of interest [see, e.g., @Zeyen:1993; @Gallardo:2003; @Moorkamp:2011].
+
+While this inversion method is limited to linear forward models, the reconstructed posterior distribution of the geophysical properties can be further refined using site-specific knowledge or more complex inversion methods for sampling non-linear forward models, such as hIPPYlib [@Hippylib], Obsidian [@Obsidian], or GemPy [@GemPy].
+
[…]
