### 2021-01-20 #2690 @sebhaan
Hi @hugoledoux,

thanks for the editorial work. I've finished the suggested points, see below inline.

> Great.
>
> At this point could you:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
Re: Release tag: v1.0.0
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
Re: Done in Zenodo
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
Re: Checked, see linke here https://zenodo.org/record/4451474
> * [ ]  Please list the DOI of the archived version here.
Re: Zenodo DOI:  10.5281/zenodo.4451474
>
> I can then move forward with accepting the submission.
