### 2021-01-22 f985bf9
Update paper.md
Fixing typos.
