### 2021-01-22 #2690 @kthyng
Hi @sebhaan I edited your paper. See if you agree with the changes in the [PR](https://github.com/sebhaan/geobo/pull/2) and merge if so.

Also, it looks like at least some words that should be capitalized in your references are not capitalized. For example, Bayesian in Brochu et al and Gaussian in Melkumyan et al. Please go through all references in detail to fix these. You can add {} around words or characters to preserve capitalization.
