### 2021-01-28 597f44b
Update capitalization  of titles in references
