+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{Melkumyan:2009,
-  title={A sparse covariance function for exact Gaussian process inference in large datasets.},
+  title={{A sparse covariance function for exact Gaussian process inference in large datasets.}},
   author={Melkumyan, Arman and Ramos, Fabio},
   booktitle={IJCAI},
   volume={9},
@@ -8,7 +8,7 @@
 }
 
 @inproceedings{Melkumyan:2011,
-  title={Multi-kernel Gaussian processes},
+  title={{Multi-kernel Gaussian processes}},
   author={Melkuyman, Armon and Ramos, Fabio},
   booktitle={IJCAI},
   volume={22},
@@ -18,7 +18,7 @@
 }
 
 @inproceedings{Reid:2013,
-  title={Bayesian Joint Inversions for the Exploration of Earth Resources.},
+  title={{Bayesian joint inversions for the exploration of earth resources.}},
   author={Reid, Alistair and Simon Timothy, O'Callaghan and Bonilla, Edwin V and McCalman, Lachlan and Rawling, Tim and Ramos, Fabio},
   booktitle={IJCAI},
   pages={2877--2884},
@@ -26,14 +26,14 @@
 }
 
 @book{Rasmussen:2006,
-  title={Gaussian process for machine learning},
+  title={{Gaussian process for machine learning}},
   author={Rasmussen, Carl Edward and Williams, Christopher KI},
   year={2006},
   publisher={MIT press}
 }
 
 @article{Li:1998,
-  title={3D-inversion of gravity data},
+  title={{3D-inversion of gravity data}},
   author={Yaoguo, Li and Oldenburg, Douglas W},
   journal={Geophysics},
   volume={63},
@@ -43,14 +43,14 @@
 }
 
 @article{Brochu:2010,
-  title={A tutorial on Bayesian optimization of expensive cost functions, with application to active user modeling and hierarchical reinforcement learning},
+  title={{A tutorial on Bayesian optimization of expensive cost functions, with application to active user modeling and hierarchical reinforcement learning}},
   author={Brochu, Eric and Cora, Vlad M and De Freitas, Nando},
   journal={arXiv preprint arXiv:1012.2599},
   year={2010}
 }
 
 @article{Haan:2020,
-  title={Multiobjective Bayesian optimisation and joint inversion for active sensor fusion},
+  title={{Multiobjective Bayesian optimisation and joint inversion for active sensor fusion}},
   author={Sebastian Haan, Fabio Ramos, Dietmar Muller},
   journal={Geophysics, 86(1), pp.1-78.},
   doi = {10.1190/geo2019-0460.1},
@@ -58,7 +58,7 @@
 }
 
 @article{Hippylib,
-  title = "{hIPPYlib: an Extensible Software Framework for Large-scale Deterministic and Bayesian Inverse Problems}",
+  title = "{{hIPPYlib: An extensible software framework for large-scale inverse problems.}}",
   author = {Villa, U. and Petra, N. and Ghattas, O.},
   journal = {Journal of Open Source Software},
   volume = {3},
@@ -70,7 +70,7 @@
 }
 
 @article{Obsidian,
-  title={Efficiency and robustness in Monte Carlo sampling for 3-D geophysical inversions with Obsidian v0. 1.2: setting up for success},
+  title={{Efficiency and robustness in Monte Carlo sampling for 3-D geophysical inversions with Obsidian v0. 1.2: setting up for success}},
   author={Scalzo, Richard and Kohn, David and Olierook, Hugo and Houseman, Gregory and Chandra, Rohitash and Girolami, Mark and Cripps, Sally},
   journal={Geoscientific Model Development},
[…]
