+++ paper/paper.md
@@ -32,7 +32,7 @@ First, ``GeoBO`` jointly solves multi-linear forward models of 2D survey data (e
 
 While this inversion method is limited to linear forward models, the reconstructed posterior distribution of the geophysical properties can be further refined using site-specific knowledge or more complex inversion methods for sampling non-linear forward models, such as hIPPYlib [@Hippylib], Obsidian [@Obsidian], or GemPy [@GemPy].
 
-In a second step, the BO in `GeoBO` allows the user to define objectives in an acquisition function [@Mockus:1975; @Brochu:2010; @Haan:2020] which typically has to balance between a) exploration, i.e., querying points that maximise the information gain and minimize the uncertainty of a geophysical site model; b) exploitation, i.e. querying points that maximise the reward (e.g., concentrating search in the vicinity locations with high value such as minerals); and c) minimize the number of samples given a cost function for any new measurement (e.g., costly drill-cores). The reconstructed 3D model output of the joint inversion model is then used to query for the next most promising point based on the aquisition function, which guides the search for a user-defined optimum.
+In a second step, the BO in `GeoBO` allows the user to define objectives in an acquisition function [@Mockus:1975; @Brochu:2010; @Haan:2020] which typically has to balance between a) exploration, i.e., querying points that maximise the information gain and minimize the uncertainty of a geophysical site model; b) exploitation, i.e., querying points that maximise the reward (e.g., concentrating search in the vicinity locations with high value such as minerals); and c) minimize the number of samples given a cost function for any new measurement (e.g., costly drill-cores). The reconstructed 3D model output of the joint inversion model is then used to query for the next most promising point based on the aquisition function, which guides the search for a user-defined optimum.
 
 
 ``GeoBO`` allows applications to specify priors as additional options, such as a) the typical correlation lengthscale of the geological structure via GP hyperparameters, b) the gain/cost parameters in the BO acquisition function, and c) the correlation coefficients between different geophysical properties. This package includes forward models for gravity and magnetics surveys [@Li:1998], drillcores, and test-sets of geological models and data. 
