### 2021-01-28 68d8f56
Minor update of replacing journal acronyms with full names
