+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{Melkumyan:2009,
   title={{A sparse covariance function for exact Gaussian process inference in large datasets.}},
   author={Melkumyan, Arman and Ramos, Fabio},
-  booktitle={IJCAI},
+  booktitle={International Joint Conference on Artificial Intelligence},
   volume={9},
   pages={1936--1942},
   year={2009}
@@ -10,7 +10,7 @@
 @inproceedings{Melkumyan:2011,
   title={{Multi-kernel Gaussian processes}},
   author={Melkuyman, Armon and Ramos, Fabio},
-  booktitle={IJCAI},
+  booktitle={International Joint Conference on Artificial Intelligence},
   volume={22},
   number={1},
   pages={1408},
@@ -20,7 +20,7 @@
 @inproceedings{Reid:2013,
   title={{Bayesian joint inversions for the exploration of earth resources.}},
   author={Reid, Alistair and Simon Timothy, O'Callaghan and Bonilla, Edwin V and McCalman, Lachlan and Rawling, Tim and Ramos, Fabio},
-  booktitle={IJCAI},
+  booktitle={International Joint Conference on Artificial Intelligence},
   pages={2877--2884},
   year={2013}
 }
