### 2021-01-28 #2690 @sebhaan
@danielskatz - .bib file updated. IJCAI replaced with International Joint Conference on Artificial Intelligence
